refactor: AuthService 생성 및 로그인 API 로직 분리
